feat: merge login function to top page
