+++ client/src/pages/@components/BasicHeader/basicHeader.module.css
@@ -17,6 +17,21 @@
   letter-spacing: 0.25em;
 }
 
+.loginBtn {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  padding: 8px 12px;
+  font-size: 14px;
+  font-weight: bold;
+  color: #fff;
+  cursor: pointer;
+  background: #fff1;
+  border: 1px solid #fff5;
+  border-radius: 4px;
+  transition: 0.2s;
+}
+
 .userBtn {
   display: flex;
   gap: 8px;
@@ -28,6 +43,7 @@
   transition: 0.2s;
 }
 
+.loginBtn:hover,
 .userBtn:hover {
   border-color: #eee;
 }
