--- client/src/pages/login/index.module.css
@@ -1,40 +1,0 @@
-.container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  backdrop-filter: blur(4px);
-}
-
-.title {
-  font-size: 80px;
-  font-weight: bold;
-  text-shadow: 1px 1px white;
-}
-
-.btn {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  padding: 8px 12px;
-  font-size: 14px;
-  font-weight: bold;
-  color: white;
-  cursor: pointer;
-  background: #222;
-  border-radius: 4px;
-  transition: 0.2s background;
-}
-
-.btn:hover {
-  background: #444;
-}
